minor: move addons out of utils
