+++ modmail/bot.py
@@ -10,9 +10,9 @@ from discord import Activity, AllowedMentions, Intents
 from discord.client import _cleanup_loop
 from discord.ext import commands
 
+from modmail.addons.plugins import PLUGINS, walk_plugins
 from modmail.config import CONFIG
 from modmail.log import ModmailLogger
-from modmail.utils.addons.plugins import PLUGINS, walk_plugins
 from modmail.utils.extensions import EXTENSIONS, NO_UNLOAD, walk_extensions
 
 REQUIRED_INTENTS = Intents(
